+++ server/SharedTestDependencies/TestObjects.cs
@@ -26,7 +26,7 @@ namespace SharedTestDependencies
                 .RuleFor(o => o.OrderEntries, new List<OrderEntry>());
         }
 
-        public static List<Paper> GetPapers(int paperCount, int propertyCount)
+        public static List<PaperApi> GetPapers(int paperCount, int propertyCount)
         {
             var papers = GeneratePapers(paperCount);
             var properties = GenerateProperties(propertyCount);
@@ -42,9 +42,9 @@ namespace SharedTestDependencies
         }
         
         // Helper to create a list of papers
-        private static List<Paper> GeneratePapers(int count)
+        private static List<PaperApi> GeneratePapers(int count)
         {
-            var paperFaker = new Faker<Paper>()
+            var paperFaker = new Faker<PaperApi>()
                 .RuleFor(p => p.Name, f => f.Commerce.ProductName())
                 .RuleFor(p => p.Discontinued, f => f.Random.Bool())
                 .RuleFor(p => p.Stock, f => f.Random.Int(0, 1000))
